Move cell to its own file
